@@ -5,17 +5,17 @@
     
     <groupId>nl.logius.osdbk</groupId>
     <artifactId>throttling-service</artifactId>
-    <version>1.10.0-SNAPSHOT</version>
+    <version>1.11.0-SNAPSHOT</version>
     <name>throttling-service</name>
     
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
     </properties>
     
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.10</version>
+        <version>3.1.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     
@@ -59,7 +59,7 @@
         <dependency>
             <groupId>net.logstash.logback</groupId>
             <artifactId>logstash-logback-encoder</artifactId>
-            <version>7.2</version>
+            <version>7.4</version>
         </dependency>
         <dependency>
             <groupId>org.codehaus.janino</groupId>
@@ -94,7 +94,7 @@
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
-                <version>0.8.6</version>
+                <version>0.8.11</version>
                 <executions>
                     <execution>
                         <id>pre-unit-test</id>
@@ -108,7 +108,7 @@
             <plugin>
                 <groupId>org.owasp</groupId>
                 <artifactId>dependency-check-maven</artifactId>
-                <version>8.1.2</version>
+                <version>8.4.2</version>
                 <configuration>
                     <suppressionFiles>
                         <suppressionFile>owasp-project-ignore.xml</suppressionFile>
